% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcAUC}
\alias{calcAUC}
\title{Fast computation of the AUC w.r.t. to the ROC}
\usage{
calcAUC(preds, y, fast = TRUE, sorted = FALSE)
}
\arguments{
\item{preds}{Numeric vector of predicted scores}

\item{y}{True binary outcomes coded as 0 or 1. Must be an integer
vector.}

\item{fast}{Shall the computation be as fast as possible?}

\item{sorted}{Are the predicted scores already sorted
increasingly? If so, this can slightly speed up the computation.}
}
\value{
The AUC between 0 and 1
}
\description{
This function computes the area under the receiver operating
characteristic curve.
}
