% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcBrier}
\alias{calcBrier}
\title{Calculate the Brier score}
\usage{
calcBrier(preds, y)
}
\arguments{
\item{preds}{Numeric vector of predictions}

\item{y}{True outcomes}
}
\value{
The Brier score
}
\description{
Computation of the Brier score, i.e., the mean squared
error for risk estimates in a binary classification
problem.
}
