% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcDev}
\alias{calcDev}
\title{Calculate the deviance}
\usage{
calcDev(preds, y)
}
\arguments{
\item{preds}{Numeric vector of predictions}

\item{y}{True outcomes}
}
\value{
The deviance
}
\description{
Computation of the deviance, i.e., two times the negative
log likelihood for risk estimates in a binary classification
problem.
}
