% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcMis}
\alias{calcMis}
\title{Calculate the misclassification rate}
\usage{
calcMis(preds, y, cutoff = 0.5)
}
\arguments{
\item{preds}{Numeric vector of predictions}

\item{y}{True outcomes}

\item{cutoff}{Classification cutoff. By default,
scores above 50% are assigned class 1 and class 0
otherwise.}
}
\value{
The misclassification rate
}
\description{
Computation of the misclassification rate for risk
estimates in a binary classification problem.
}
