% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcNRMSE}
\alias{calcNRMSE}
\title{Calculate the NRMSE}
\usage{
calcNRMSE(preds, y, type = "sd")
}
\arguments{
\item{preds}{Numeric vector of predictions}

\item{y}{True outcomes}

\item{type}{\code{"sd"} uses the standard deviation of \code{y} for
normalization. \code{"range"} uses the whole span of \code{y}.}
}
\value{
The NRMSE
}
\description{
Computation of the normalized root mean squared error.
}
