% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vim}
\alias{plot.vim}
\title{Plot calculated VIMs}
\usage{
\method{plot}{vim}(x, p = 10, ...)
}
\arguments{
\item{x}{An object of the class \code{vim}}

\item{p}{The number of most important terms which will be included
in the plot. A value of 0 leads to plotting all terms.}

\item{...}{Ignored additional parameters}
}
\value{
No return value, called for side effects
}
\description{
This function plots variable importance measures yielded by the function
\code{\link{vim}} in a dotchart.
}
