% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitSNPs}
\alias{splitSNPs}
\title{Split biallelic SNPs into binary variables}
\usage{
splitSNPs(data)
}
\arguments{
\item{data}{A matrix or data frame only consisting of SNPs to
be splitted}
}
\value{
A data frame of the splitted SNPs
}
\description{
This function takes a matrix or data frame of SNPs coded as
0, 1, 2 or 1, 2, 3 and returns a data frame with twice as many
columns. SNPs are splitted into dominant and recessive modes,
i.e., for a \eqn{\mathrm{SNP} \in \lbrace 0,1,2 \rbrace}, two variables
\eqn{\mathrm{SNP}_D = (\mathrm{SNP} \neq 0)} and
\eqn{\mathrm{SNP}_R = (\mathrm{SNP} = 2)} are generated.
}
