% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcMSE}
\alias{calcMSE}
\title{Calculate the MSE}
\usage{
calcMSE(preds, y)
}
\arguments{
\item{preds}{Numeric vector of predictions}

\item{y}{True outcomes}
}
\value{
The MSE
}
\description{
Computation of the mean squared error.
}
