% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.boosting.R
\name{fitLinearBoostingModel}
\alias{fitLinearBoostingModel}
\title{Linear models based on boosted models}
\usage{
fitLinearBoostingModel(model, n.iter)
}
\arguments{
\item{model}{Fitted \code{logic.boosted} model}

\item{n.iter}{Number of boosting iterations to be used}
}
\value{
A \code{linear.logic} model. This is a list containing
  the logic terms used as predictors in the model and the fitted \code{glm}
  model.
}
\description{
This function uses a fitted \code{logic.boosted} model for fitting
a linear or logistic (depending on the type of outcome) regression model.
}
\details{
In this procedure, the logic terms are extracted from the individual
\code{logicDT} models and the set of unique terms are used as predictors
in a regression model. For incorporating a continuous covariable
the covariable itself as well as products of the covariable with the
extracted logic terms are included as predictors in the regression model.
}
