% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicDT.R
\name{predict.4pl}
\alias{predict.4pl}
\title{Prediction for 4pL models}
\usage{
\method{predict}{`4pl`}(object, Z, ...)
}
\arguments{
\item{object}{Fitted 4pl model}

\item{Z}{Numeric vector of new input samples}

\item{...}{Ignored additional parameters}
}
\value{
A numeric vector of predictions. For binary outcomes,
  this is a vector with estimates for
  \eqn{P(Y=1 \mid X = x)}.
}
\description{
Use new input data and a fitted four parameter logistic
model to predict corresponding outcomes.
}
