% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicDT.R
\name{predict.linear}
\alias{predict.linear}
\title{Prediction for linear models}
\usage{
\method{predict}{linear}(object, Z, ...)
}
\arguments{
\item{object}{Fitted \code{linear} model}

\item{Z}{Numeric vector of new input samples}

\item{...}{Ignored additional parameters}
}
\value{
A numeric vector of predictions. For binary outcomes,
  this is a vector with estimates for
  \eqn{P(Y=1 \mid X = x)}.
}
\description{
Use new input data and a fitted linear
model to predict corresponding outcomes.
}
\details{
For binary outcomes, predictions are cut at 0 or 1 for generating
proper probability estimates.
}
