% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicDT.R
\name{refitTrees}
\alias{refitTrees}
\title{Refit the logic decision trees}
\usage{
refitTrees(model, tree_control)
}
\arguments{
\item{model}{A fitted \code{logicDT} model}

\item{tree_control}{Tree control parameters. This object should be
constructed using the function \code{\link{tree.control}}.
Alternatively, the old \code{tree_control} from \code{model} can be
modified and specified here.}
}
\value{
The \code{logicDT} model with newly fitted trees
}
\description{
Newly fit the decision trees in the \code{logicDT} model using
the supplied tree control parameters.
This is especially useful if, e.g., the model was initially trained
without utilizing a continuous covariable or fitting linear models and
now 4pL model shall be fitted.
}
